+++ pycfmodel/model/resources/iam_user.py
@@ -5,6 +5,7 @@ from pycfmodel.model.parameter import Parameter
 from pycfmodel.model.resources.properties.policy import Policy
 from pycfmodel.model.resources.resource import Resource
 from pycfmodel.model.types import Resolvable, ResolvableStr
+from pycfmodel.model.utils import OptionallyNamedPolicyDocument
 
 
 class IAMUserProperties(CustomModel):
@@ -45,7 +46,7 @@ class IAMUser(Resource):
     Properties: Optional[Resolvable[IAMUserProperties]]
 
     def has_hardcoded_credentials(self) -> bool:
-        """ Returns True if login profile password contains a hardcoded string, otherwise False. """
+        """Returns True if login profile password contains a hardcoded string, otherwise False."""
         if self.Properties:
             login_profile = self.Properties.LoginProfile
             if login_profile and login_profile.get("Password"):
@@ -53,3 +54,11 @@ class IAMUser(Resource):
                     return True
 
         return super().has_hardcoded_credentials()
+
+    @property
+    def policy_documents(self) -> List[OptionallyNamedPolicyDocument]:
+        result = []
+        policies = self.Properties.Policies if self.Properties and self.Properties.Policies else []
+        for policy in policies:
+            result.append(OptionallyNamedPolicyDocument(policy.PolicyName, policy.PolicyDocument))
+        return result