+++ pycfmodel/model/resources/sqs_queue_policy.py
@@ -4,6 +4,7 @@ from pycfmodel.model.base import CustomModel
 from pycfmodel.model.resources.properties.policy_document import PolicyDocument
 from pycfmodel.model.resources.resource import Resource
 from pycfmodel.model.types import Resolvable, ResolvableStr
+from pycfmodel.model.utils import OptionallyNamedPolicyDocument
 
 
 class SQSQueuePolicyProperties(CustomModel):
@@ -32,3 +33,7 @@ class SQSQueuePolicy(Resource):
     TYPE_VALUE: ClassVar = "AWS::SQS::QueuePolicy"
     Type: str = TYPE_VALUE
     Properties: Resolvable[SQSQueuePolicyProperties]
+
+    @property
+    def policy_documents(self) -> List[OptionallyNamedPolicyDocument]:
+        return [OptionallyNamedPolicyDocument(name=None, policy_document=self.Properties.PolicyDocument)]