+++ tests/resources/properties/test_policy_document.py
@@ -1,13 +1,13 @@
 import re
-from collections import Counter
 
-import pytest
+from pytest import fixture
 
+from pycfmodel.cloudformation_actions import CLOUDFORMATION_ACTIONS
 from pycfmodel.constants import CONTAINS_STAR
 from pycfmodel.model.resources.properties.policy_document import PolicyDocument
 
 
-@pytest.fixture()
+@fixture
 def policy_document_one_statement():
     return PolicyDocument(
         **{
@@ -21,7 +21,7 @@ def policy_document_one_statement():
     )
 
 
-@pytest.fixture()
+@fixture
 def policy_document_multi_statement():
     return PolicyDocument(
         **{
@@ -42,14 +42,14 @@ def policy_document_multi_statement():
     )
 
 
-@pytest.fixture()
+@fixture
 def policy_document_star_resource():
     return PolicyDocument(
         **{"Statement": [{"Action": ["*"], "Effect": "Allow", "Resource": "*", "Principal": {"AWS": ["156460612806"]}}]}
     )
 
 
-@pytest.fixture()
+@fixture
 def policy_document_wildcard_actions():
     return PolicyDocument(
         **{
@@ -65,17 +65,28 @@ def policy_document_wildcard_actions():
     )
 
 
-@pytest.fixture()
+@fixture
 def policy_document_not_principal():
     return PolicyDocument(
         **{
             "Statement": [
                 {
-                    "Action": ["IAM:Delete*"],
+                    "Action": [
+                        "iam:Delete*",
+                        "s3:GetObject*",
+                    ],
                     "Effect": "Allow",
                     "Resource": "arn:aws:s3:::fakebucketfakebucket/*",
                     "NotPrincipal": {"AWS": ["156460612806"]},
-                }
+                },
+                {
+                    "Action": [
+                        "s3:List*",
+                    ],
+                    "Effect": "Deny",
+                    "Resource": "arn:aws:s3:::fakebucketfakebucket/*",
+                    "NotPrincipal": {"AWS": ["156460612806"]},
+                },
             ]
         }
     )
@@ -87,7 +98,7 @@ def test_one_statement(policy_document_one_statement):
 
 def test_multi_statements(policy_document_multi_statement):
     assert policy_document_multi_statement.Statement[0].Effect == "Allow"
-    assert policy_document_multi_statement.Statement[1].Effect == "bar"
+    assert policy_document_multi_statement.Statement[1].Effect == "Bar"
 
 
 def test_star_resource(policy_document_star_resource):
@@ -117,27 +128,93 @@ def test_non_whitelisted_allowed_principals(policy_document_multi_statement):
 
 def test_get_iam_actions(policy_document_not_principal):
     correct_list = [
-        "IAM:DeleteAccountPasswordPolicy",
-        "IAM:DeleteServiceLinkedRole",
-        "IAM:DeleteRole",
-        "IAM:DeleteOpenIDConnectProvider",
-        "IAM:DeleteGroup",
-        "IAM:DeleteRolePolicy",
-        "IAM:DeleteSSHPublicKey",
-        "IAM:DeleteLoginProfile",
-        "IAM:DeleteServiceSpecificCredential",
-        "IAM:DeleteUserPolicy",
-        "IAM:DeleteVirtualMFADevice",
-        "IAM:DeletePolicyVersion",
-        "IAM:DeleteGroupPolicy",
-        "IAM:DeleteAccountAlias",
-        "IAM:DeleteSigningCertificate",
-        "IAM:DeleteUser",
-        "IAM:DeletePolicy",
-        "IAM:DeleteSAMLProvider",
-        "IAM:DeleteAccessKey",
-        "IAM:DeleteServerCertificate",
-        "IAM:DeleteInstanceProfile",
+        "iam:DeleteAccessKey",
+        "iam:DeleteAccountAlias",
+        "iam:DeleteAccountPasswordPolicy",
+        "iam:DeleteGroup",
+        "iam:DeleteGroupPolicy",
+        "iam:DeleteInstanceProfile",
+        "iam:DeleteLoginProfile",
+        "iam:DeleteOpenIDConnectProvider",
+        "iam:DeletePolicy",
+        "iam:DeletePolicyVersion",
+        "iam:DeleteRole",
+        "iam:DeleteRolePermissionsBoundary",
+        "iam:DeleteRolePolicy",
+        "iam:DeleteSAMLProvider",
+        "iam:DeleteSSHPublicKey",
+        "iam:DeleteServerCertificate",
+        "iam:DeleteServiceLinkedRole",
+        "iam:DeleteServiceSpecificCredential",
+        "iam:DeleteSigningCertificate",
+        "iam:DeleteUser",
+        "iam:DeleteUserPermissionsBoundary",
+        "iam:DeleteUserPolicy",
+        "iam:DeleteVirtualMFADevice",
     ]
 
-    assert Counter(correct_list) == Counter(policy_document_not_principal.get_iam_actions())
+    assert policy_document_not_principal.get_iam_actions() == correct_list
+
+
+def test_get_allowed_actions(policy_document_not_principal):
+    correct_list = [
+        "iam:DeleteAccessKey",
+        "iam:DeleteAccountAlias",
+        "iam:DeleteAccountPasswordPolicy",
+        "iam:DeleteGroup",
+        "iam:DeleteGroupPolicy",
+        "iam:DeleteInstanceProfile",
+        "iam:DeleteLoginProfile",
+        "iam:DeleteOpenIDConnectProvider",
+        "iam:DeletePolicy",
+        "iam:DeletePolicyVersion",
+        "iam:DeleteRole",
+        "iam:DeleteRolePermissionsBoundary",
+        "iam:DeleteRolePolicy",
+        "iam:DeleteSAMLProvider",
+        "iam:DeleteSSHPublicKey",
+        "iam:DeleteServerCertificate",
+        "iam:DeleteServiceLinkedRole",
+        "iam:DeleteServiceSpecificCredential",
+        "iam:DeleteSigningCertificate",
+        "iam:DeleteUser",
+        "iam:DeleteUserPermissionsBoundary",
+        "iam:DeleteUserPolicy",
+        "iam:DeleteVirtualMFADevice",
+        "s3:GetObject",
+        "s3:GetObjectAcl",
+        "s3:GetObjectLegalHold",
+        "s3:GetObjectRetention",
+        "s3:GetObjectTagging",
+        "s3:GetObjectTorrent",
+        "s3:GetObjectVersion",
+        "s3:GetObjectVersionAcl",
+        "s3:GetObjectVersionForReplication",
+        "s3:GetObjectVersionTagging",
+        "s3:GetObjectVersionTorrent",
+    ]
+
+    assert policy_document_not_principal.get_allowed_actions() == correct_list
+
+
+@fixture
+def policy_document_not_action():
+    return PolicyDocument(
+        **{
+            "Statement": [
+                {
+                    "NotAction": [
+                        "rds:*",
+                    ],
+                    "Effect": "Allow",
+                    "Resource": "arn:aws:s3:::fakebucketfakebucket/*",
+                    "NotPrincipal": {"AWS": ["156460612806"]},
+                },
+            ]
+        }
+    )
+
+
+def test_get_allowed_not_actions(policy_document_not_action):
+    allowed_actions = set(entry for entry in CLOUDFORMATION_ACTIONS if not entry.startswith("rds:"))
+    assert policy_document_not_action.get_allowed_actions() == sorted(allowed_actions)