+++ tests/resources/test_s3_bucket_policy.py
@@ -1,6 +1,7 @@
 import pytest
 
 from pycfmodel.model.resources.s3_bucket_policy import S3BucketPolicy
+from pycfmodel.model.utils import OptionallyNamedPolicyDocument
 
 
 @pytest.fixture()
@@ -27,3 +28,9 @@ def s3_bucket_policy():
 
 def test_policy_document(s3_bucket_policy):
     assert s3_bucket_policy.Properties.PolicyDocument.Statement[0].Action[0] == "*"
+
+
+def test_s3_bucketpolicy_policy_document_property(s3_bucket_policy):
+    assert s3_bucket_policy.policy_documents == [
+        OptionallyNamedPolicyDocument(name=None, policy_document=s3_bucket_policy.Properties.PolicyDocument)
+    ]